// net/js/enterprise-nx6-workspace/libs/commerce/order-lib/src/api/model/integrationEventLogEntry.ts
/**
 * Ordering HTTP API
 * The Ordering Service HTTP API
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IntegrationEventLogEntry {
  eventId?: string;
  eventTypeName?: string;
  state?: IntegrationEventLogEntry.StateEnum;
  timesSent?: number;
  creationTime?: Date;
  content?: string;
  orderId?: number;
  orderStatus?: string;
}
export namespace IntegrationEventLogEntry {
  export type StateEnum = 'NotPublished' | 'Published' | 'PublishedFailed';
  export const StateEnum = {
    NotPublished: 'NotPublished' as StateEnum,
    Published: 'Published' as StateEnum,
    PublishedFailed: 'PublishedFailed' as StateEnum
  };
}